consistency

+++ code/math/sieve.cpp
@@ -4,15 +4,18 @@
 //
 // Sieve[i] == true if i is prime
 //
-// build_sieve - O(n log(log(n)))
+// build_sieve - O(n.log(log(n)))
 
 vector<bool> sieve(MAX,true);
 
 void build_sieve(int lim) {
 	sieve[0] = sieve[1] = false;
-	for(ll i=4; i < lim; i += 2) sieve[i] = false;
-	for(ll i=3; i < lim; i += 2) if(sieve[i] == true)
-		for(ll j = i*i; j < lim; j += i) sieve[j] = false;
+	for (int i = 4; i < lim; i += 2) 
+		sieve[i] = false;
+	for (int i = 3; i < lim; i += 2) 
+		if (sieve[i] == true)
+			for (ll j = ll(i) * i; j < lim; j += i) 
+				sieve[j] = false;
 }
 
 // Sieve of Divisors
@@ -28,12 +31,16 @@ int sieve[MAX];
 void build_sieve(int lim) {
 	fill(sieve, sieve + lim, 1);
 
-	for(int i=2; i < lim; i += 2) sieve[i] = 2;
-	for(int i=3; i < lim; i += 2) if(sieve[i] == 1)
-		for(int j = i; j < lim; j += i) sieve[j] = i;
+	for (int i = 2; i < lim; i += 2) 
+		sieve[i] = 2;
+	for (int i = 3; i < lim; i += 2) 
+		if (sieve[i] == 1)
+			for (int j = i; j < lim; j += i) 
+				sieve[j] = i;
 }
 
 void fact(vector<int>& v, int n) {
-	if (n != sieve[n]) fact(v, n/sieve[n]);
+	if (n != sieve[n]) 
+		fact(v, n / sieve[n]);
 	v.push_back(sieve[n]);
 }
